third_party/blink/web_tests/fast/forms/range/range-keyboard-oninput-event-expected.txt

Test that for each keyboard event both input and change event are dispatched

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Move left
PASS inputEventCounter is >= lastInputEventCounter + 1
PASS changeEventCounter is >= lastChangeEventCounter + 1
Move right
PASS inputEventCounter is >= lastInputEventCounter + 1
PASS changeEventCounter is >= lastChangeEventCounter + 1
PASS successfullyParsed is true

TEST COMPLETE
